refactor: 지출(Expense) 관련 데이터 및 도메인 레이어 패키지 구조 개선
지출 관련 코드의 응집도를 높이고 패키지 구조를 체계화하기 위해, 기존 `diary` 패키지에 흩어져 있던 지출 관련 파일들을 독립된 `expense` 패키지로 이동하고 관련 의존성을 업데이트합니다.

**주요 변경 사항:**

*   **패키지 구조 재구성 및 파일 이동:**
    *   **Data 레이어**: `data/diary` 하위의 `ExpenseDao`, `ExpenseEntity`, `RoomExpenseRepository`를 각각 `data/expense/dao`, `data/expense/entity`, `data/expense/repository`로 이동했습니다.
    *   **Domain 레이어**: `domain/repository/ExpenseRepository`를 `domain/expense/ExpenseRepository`로 이동했습니다.
*   **클래스 및 인터페이스 수정:**
    *   `RoomExpenseRepository`를 `ExpenseRepositoryImpl`로 이름을 변경하고, Hilt 의존성 주입을 위해 `@Inject constructor`를 추가했습니다.
    *   `ExpenseEntity`의 외래 키(FK) 관련 주석을 다듬고 가독성을 개선했습니다.
*   **의존성 및 임포트 업데이트:**
    *   파일 이동에 따라 `PicDayDatabase`, `DiaryModule`, 그리고 지출 관련 UseCase들(`AddExpenseUseCase`, `DeleteExpenseUseCase`, `GetDiaryExpensesUseCase` 등)의 임포트 경로를 수정했습니다.
    *   `DiaryModule`에서 `ExpenseRepository` 제공 시 `ExpenseRepositoryImpl`을 사용하도록 변경했습니다.
