+++ app/src/main/java/com/picday/diary/domain/usecase/expense/GetDiaryExpenseTotalUseCase.kt
@@ -1,6 +1,6 @@
 package com.picday.diary.domain.usecase.expense
 
-import com.picday.diary.domain.repository.ExpenseRepository
+import com.picday.diary.domain.expense.ExpenseRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
